dev/data-science-project/requirements.txt
torch==1.12.1
torchvision==0.13.1
easyocr==1.4.1
opencv-python-headless==4.5.3.20210927
ultralytics==8.0.0
pandas==1.3.3
numpy==1.21.2
PyYAML==5.4.1
matplotlib==3.4.3
shutil==1.0.0